feat: move api to follow second entry point structure
