+++ lib/core/api/README.md
@@ -1,0 +1,3 @@
+# @alfresco/adf-core/api
+
+Secondary entry point of `@alfresco/adf-core`. It can be used by importing from `@alfresco/adf-core/api`.
