+++ lib/core/api/ng-package.json
@@ -1,6 +1,5 @@
 {
-  "$schema": "../../../../../node_modules/ng-packagr/ng-package.schema.json",
   "lib": {
-    "entryFile": "public-api.ts"
+    "entryFile": "src/index.ts"
   }
 }
