+++ lib/core/api/src/index.ts
@@ -15,7 +15,8 @@
  * limitations under the License.
  */
 
-export * from './activiti/activiti-client.types';
-export * from './alfresco-js-clients.module';
-export * from './discovery/discovery-client.types';
-
+export * from './lib/api-client.factory';
+export * from './lib/api-clients.service';
+export * from './lib/clients';
+export * from './lib/types';
+export * from './lib/alfresco-api/alfresco-api.http-client';
