+++ lib/core/api/src/lib/alfresco-api/alfresco-api.http-client.spec.ts
