+++ lib/core/api/src/lib/clients/activiti/activiti-client.types.ts
