+++ lib/core/api/src/lib/clients/discovery/discovery-client.module.ts
