+++ lib/core/api/src/lib/clients/discovery/discovery-client.types.ts
