+++ lib/core/api/src/lib/clients/index.ts
@@ -15,8 +15,6 @@
  * limitations under the License.
  */
 
-export * from './api-client.factory';
-export * from './api-clients.service';
-export * from './clients';
-export * from './types';
-export * from './alfresco-api/alfresco-api.http-client';
+export * from './activiti/activiti-client.types';
+export * from './alfresco-js-clients.module';
+export * from './discovery/discovery-client.types';
